+++ ui/PrintButton.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 /* global window */
 import Button from 'react-bootstrap/Button';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
